Rename definition files to match module file names
The files in the module were renamed.
