+++ types/socketcluster-client/lib/clientsocket.d.ts
